# Send Instance health changed event when the overall status changed