Before this fix, during a deployment triggered by a restart, the old
instances of an app were killed before the new ones were healthy. This
was due to the "instance_health_changed_event" event sent too early to
the ReadinessBehavior instance (when the first health check reported to be
healthy instead of ALL health checks).

The aim of this fix is to aggregate the results of health checks and
maintain a global instance health state which is Some(true) if all health
checks of a given instances report healthy, Some(false) if all health checks
report unhealthy and Option.empty if there is at least one unreported result.
The "instance_health_changed_event" event is now sent only when there is a
transition between those 3 states.

This commit relates to those tickets: MARATHON-7609, MARATHON-2683.

Backport in branch v1.5: https://github.com/mesosphere/marathon/pull/6045