@pierresouchay 2018-02-23
This is a really important change for us as we have apps that have several healthchecks and some of them take time to be up, especially for applications with a warming up phase.

It really cause us some troubles on some of our apps, and it would benefits to many to have this kind of sane behaviour, especially when restarting applications or upgrading applications with new versions.