@clems4ever 2018-03-06
Hello @alenkacz, no problem. It has been a pain in our company for the restart of apps, thus we decided to tackle it.
Regarding this PR, I simplified a bit the code on our end. I'm rebasing right now.