@kensipe 2018-03-06
@clems4ever apologies... we have it setup so that PRs require a committer to bump the build.   this prevents malicious executions... if everyone is paying attention... let me help with that