@clems4ever 2018-03-07
Thank you very much for the review @kensipe! This one was really critical for us since we have more and more apps deployed in our platform. I guess a lot of users will be happy to have it in the next release :).