@clems4ever 2018-03-08
@kensipe , I added some documentation and updated the logs.