@clems4ever 2018-03-10
@timcharper , nice catch! Thank you for the review.
@kensipe , I think there's an easy way to avoid the leak, I will implement it.

Regarding the benchmark, do you have a some sort of test suite you'd like me to use or should I build one from scratch?