@alenkacz 2018-03-12
@clems4ever see this part of the repo https://github.com/mesosphere/marathon/tree/master/benchmark if you need any help with that, ping me on slack :-)