@clems4ever 2018-03-14
I've generated the benchmarks but not have the time to analyze yet.

![mesosphere marathon core health impl apphealthcheckbenchmark addhealthchecks-averagetime-2 jfr](https://user-images.githubusercontent.com/3193257/37394618-7dcd9b34-2774-11e8-91a0-6e5a3513d8fd.png)
![mesosphere marathon core health impl apphealthcheckbenchmark addhealthchecks-throughput-0 jfr](https://user-images.githubusercontent.com/3193257/37394619-7de865c2-2774-11e8-91d1-139856e92029.png)
![mesosphere marathon core health impl apphealthcheckbenchmark updatehealthcheckstatuses-averagetime-3 jfr](https://user-images.githubusercontent.com/3193257/37394620-7e2bde10-2774-11e8-9b81-5b52a0c922d9.png)
![mesosphere marathon core health impl apphealthcheckbenchmark updatehealthcheckstatuses-throughput-1 jfr](https://user-images.githubusercontent.com/3193257/37394621-7e42d2aa-2774-11e8-9ab7-1acbdda94042.png)