@clems4ever 2018-03-16
Hey, I've removed the graphs since they are not really useful here. I've done some optimizations and benchmarked them. Here are the the two tests:
1/ 1000 apps having 3 versions and 3 health checks. (addHealthCheck)
2/ 10000 randomly selected updates of health check statuses in bulk (updateHealthCheckStatus)

And the results
```
*Without optim*

[info] # Run complete. Total time: 00:02:42
[info] Benchmark                                           Mode  Cnt  Score   Error   Units
[info] AppHealthCheckBenchmark.addHealthChecks            thrpt   20  0.606 ± 0.005  ops/ms
[info] AppHealthCheckBenchmark.updateHealthCheckStatuses  thrpt   20  0.118 ± 0.004  ops/ms
[info] AppHealthCheckBenchmark.addHealthChecks             avgt   20  1.680 ± 0.017   ms/op
[info] AppHealthCheckBenchmark.updateHealthCheckStatuses   avgt   20  9.708 ± 0.594   ms/op

*With optim and immutable map*

[info] # Run complete. Total time: 00:02:42
[info] Benchmark                                           Mode  Cnt  Score   Error   Units
[info] AppHealthCheckBenchmark.addHealthChecks            thrpt   20  0.590 ± 0.017  ops/ms
[info] AppHealthCheckBenchmark.updateHealthCheckStatuses  thrpt   20  0.118 ± 0.003  ops/ms
[info] AppHealthCheckBenchmark.addHealthChecks             avgt   20  1.727 ± 0.032   ms/op
[info] AppHealthCheckBenchmark.updateHealthCheckStatuses   avgt   20  8.528 ± 0.214   ms/op


*With optim and mutable map*

[info] # Run complete. Total time: 00:02:42
[info] Benchmark                                           Mode  Cnt  Score   Error   Units
[info] AppHealthCheckBenchmark.addHealthChecks            thrpt   20  0.738 ± 0.019  ops/ms
[info] AppHealthCheckBenchmark.updateHealthCheckStatuses  thrpt   20  0.143 ± 0.027  ops/ms
[info] AppHealthCheckBenchmark.addHealthChecks             avgt   20  1.285 ± 0.009   ms/op
[info] AppHealthCheckBenchmark.updateHealthCheckStatuses   avgt   20  4.985 ± 0.118   ms/op
```

Here is my configuration:  8 core CPU and 32GB of RAM

Architecture:          x86_64
CPU op-mode(s):        32-bit, 64-bit
Byte Order:            Little Endian
CPU(s):                8
On-line CPU(s) list:   0-7
Thread(s) per core:    2
Core(s) per socket:    4
Socket(s):             1
NUMA node(s):          1
Vendor ID:             GenuineIntel
CPU family:            6
Model:                 94
Model name:            Intel(R) Xeon(R) CPU E3-1270 v5 @ 3.60GHz
Stepping:              3
CPU MHz:               3600.000
CPU max MHz:           4000,0000
CPU min MHz:           800,0000
BogoMIPS:              7200.00
Virtualization:        VT-x
L1d cache:             32K
L1i cache:             32K
L2 cache:              256K
L3 cache:              8192K
NUMA node0 CPU(s):     0-7
Flags:                 fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush dts acpi mmx fxsr sse sse2 ss ht tm pbe syscall nx pdpe1gb rdtscp lm constant_tsc art arch_perfmon pebs bts rep_good nopl xtopology nonstop_tsc cpuid aperfmperf tsc_known_freq pni pclmulqdq dtes64 monitor ds_cpl vmx smx est tm2 ssse3 sdbg fma cx16 xtpr pdcm pcid sse4_1 sse4_2 x2apic movbe popcnt tsc_deadline_timer aes xsave avx f16c rdrand lahf_lm abm 3dnowprefetch cpuid_fault epb invpcid_single pti intel_pt tpr_shadow vnmi flexpriority ept vpid fsgsbase tsc_adjust bmi1 hle avx2 smep bmi2 erms invpcid rtm mpx rdseed adx smap clflushopt xsaveopt xsavec xgetbv1 xsaves dtherm ida arat pln pts hwp hwp_notify hwp_act_window hwp_epp

And 2 slots of RAM with the following specs:

Handle 0x000A, DMI type 17, 40 bytes
Memory Device
	Array Handle: 0x0009
	Error Information Handle: Not Provided
	Total Width: 64 bits
	Data Width: 64 bits
	Size: 16384 MB
	Form Factor: DIMM
	Set: None
	Locator: DIMM1
	Bank Locator: Not Specified
	Type: DDR4
	Type Detail: Synchronous Unbuffered (Unregistered)
	Speed: 2400 MHz
	Manufacturer: 01980000802C
	Serial Number: E8391463
	Asset Tag: 04171700
	Part Number: KYXC0V-MIB          
	Rank: 2
	Configured Clock Speed: 2133 MHz
	Minimum Voltage: Unknown
	Maximum Voltage: Unknown
	Configured Voltage: 1.2 V