@clems4ever 2018-03-19
Can someone tell me why the tests end red at head since I cannot check the logs?