@timcharper 2018-03-26
Hey @clems4ever , I'll get the tests running today and will investigate this, sorry for the delay! It has taken me a while to go through all of my notifications.