@timcharper 2018-03-27
Jenkins cannot run the tests because the following file conflicts:

- src/main/scala/mesosphere/marathon/core/health/impl/MarathonHealthCheckManager.scala

I will try and watch more carefully. Feel free to flag us in the [Marathon community slack channel](https://mesos.slack.com/messages/C1L7D22KY) if we miss the build approval notification.