@clems4ever 2018-03-27
I rebased the PR after the last commit.