@clems4ever 2018-03-28
I've added more documentation as suggested.